Change how libssh2 is vendored
In the past, libssh2 had to be copied into its sub-directory (with
strange nesting). This change instead uses git submodules to make
maintenance easier.
I also added a new dev directory with a single script (for now) that
should make it easier to keep the submodule and Cython code updated.